refactor the home page
