+++ src/App.js
@@ -6,8 +6,8 @@ import React from "react";
 
 import Header from "./components/Header"
 import Navigation from "./components/Navigation";
-import Reviews from "./components/Reviews"
-import Home from "./components/Home"
+import Reviews from "./components/Home"
+import Home from "./components/Reviews"
 import Review from "./components/Review";
 
 const App = () => {
@@ -17,8 +17,8 @@ const App = () => {
         <Header />
         <Navigation />
         <Routes>
-          <Route path="/" element={<Home/>} />
-          <Route path="/reviews" element={<Reviews/>} />
+          <Route path="/reviews" element={<Home/>} />
+          <Route path="/" element={<Reviews/>} />
           <Route path="/reviews/:review_id" element={<Review/>} />
         </Routes>
       </div>
